PartPropertySidebarコンポーネントのNumberInputのラップをFragmentから直接のキー指定に変更し、usePartReducerフックをリファクタリングして、パーツ操作のロジックを一元管理しました。また、不要なpartReducerファイルを削除しました。これにより、コードの可読性と保守性が向上しました。
